requirements.txt
openai
flask
websockets
adafruit-blinka
adafruit-circuitpython-dotstar
pyaudio
setuptools
gpiozero
pygame
python-dotenv
git+https://github.com/yeutterg/openai_realtime_client_headless#egg=openai_realtime_client